iox-#32: moved platform types
