+++ iceoryx_posh/include/iceoryx_posh/internal/roudi/roudi_lock.hpp
@@ -16,11 +16,11 @@
 
 #include "iceoryx_posh/internal/log/posh_logging.hpp"
 #include "iceoryx_utils/cxx/smart_c.hpp"
+#include "iceoryx_utils/platform/types.hpp"
 
 #include <arpa/inet.h>
 #include <stdlib.h>
 #include <sys/socket.h>
-#include <sys/types.h>
 #include <unistd.h>
 
 namespace iox
