+++ iceoryx_posh/source/roudi/roudi_lock.cpp
@@ -17,7 +17,6 @@
 #include <arpa/inet.h>
 #include <stdlib.h>
 #include <sys/socket.h>
-#include <sys/types.h>
 #include <unistd.h>
 
 namespace iox
