+++ iceoryx_utils/platform/linux/include/iceoryx_utils/platform/resource.hpp
@@ -14,4 +14,4 @@
 
 #pragma once
 
-#include <semaphore.h>
+#include
